Addition of BooleanSerde
