+++ streams/src/main/java/org/apache/kafka/streams/state/internals/NullableValueAndTimestampSerde.java
@@ -48,42 +48,4 @@ public class NullableValueAndTimestampSerde<V> extends WrappingNullableSerde<Val
         );
     }
 
-    static final class BooleanSerde {
-        private static final byte TRUE = 0x01;
-        private static final byte FALSE = 0x00;
-
-        static class BooleanSerializer implements Serializer<Boolean> {
-            @Override
-            public byte[] serialize(final String topic, final Boolean data) {
-                if (data == null) {
-                    return null;
-                }
-
-                return new byte[] {
-                    data ? TRUE : FALSE
-                };
-            }
-        }
-
-        static class BooleanDeserializer implements Deserializer<Boolean> {
-            @Override
-            public Boolean deserialize(final String topic, final byte[] data) {
-                if (data == null) {
-                    return null;
-                }
-
-                if (data.length != 1) {
-                    throw new SerializationException("Size of data received by BooleanDeserializer is not 1");
-                }
-
-                if (data[0] == TRUE) {
-                    return true;
-                } else if (data[0] == FALSE) {
-                    return false;
-                } else {
-                    throw new SerializationException("Unexpected byte received by BooleanDeserializer: " + data[0]);
-                }
-            }
-        }
-    }
 }
